pyproject.toml
[tool.poetry]
name = "django-temp"
version = "0.1.0"
description = ""
authors = ["Павел <78384513+piligrimdev@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12"
django = "5.1.4"
django-rest-framework = "0.1.0"
psycopg2-binary = "2.9.10"
celery = "5.4.0"
pika = "1.3.2"
python-dotenv = "1.0.1"
gunicorn = "23.0.0"
flake8 = "7.1.1"
pep8-naming = "0.14.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
